requirements.txt
fastapi==0.135.1
uvicorn==0.34.0
httpx==0.28.1
aiosqlite==0.22.1
asyncpg==0.30.0
websockets==16.0
pydantic==2.12.5
python-dotenv==1.2.2
